Django/APIProducts/requirements.txt
django==4.2.4
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
datetime
drf-yasg
django-rest-swagger
